+++ src/layman/common/micka/__init__.py
@@ -1,0 +1,11 @@
+PROPERTIES = {
+    'md_file_identifier': {
+        'xpath': '/gmd:MD_Metadata/gmd:fileIdentifier/gco:CharacterString/text()',
+    },
+    'md_date_stamp': {
+        'xpath': '/gmd:MD_Metadata/gmd:dateStamp/gco:Date/text()',
+    },
+    'reference_system': {
+        'xpath': '/gmd:MD_Metadata/gmd:dateStamp/gco:Date/text()',
+    },
+}